+++ test/ThreadSafeQueueTest.cpp
@@ -153,7 +153,7 @@ TEST(ThreadSafeQueue, Concurrency) {
       ql::ranges::sort(result);
     }
     EXPECT_THAT(result, ::testing::ElementsAreArray(std::views::iota(
-                            uint64_t{0}, numValues * numThreads)));
+                            size_t{0}, numValues * numThreads)));
   };
   runWithBothQueueTypes(runTest);
 }
@@ -403,7 +403,7 @@ struct RunQueueManagerTest {
         ql::ranges::sort(result);
       }
       EXPECT_THAT(result, ::testing::ElementsAreArray(
-                              std::views::iota(uint64_t{0}, numValues)));
+                              std::views::iota(size_t{0}, numValues)));
     }
     // The probably most important test of all is that the destructors which are
     // run at the following closing brace never lead to a deadlock.
