+++ test/engine/OptionalJoinTest.cpp
@@ -373,7 +373,7 @@ TEST(OptionalJoin, gallopingJoin) {
       bInput.emplace_back(std::vector<IntOrId>{i, i + 12});
     }
     auto numElementsInLarger = static_cast<int64_t>(
-        std::max(uint64_t{10000}, a.numRows() * GALLOP_THRESHOLD + 1));
+        std::max(size_t{10000}, a.numRows() * GALLOP_THRESHOLD + 1));
     for (int64_t i = 400; i < numElementsInLarger; ++i) {
       bInput.emplace_back(std::vector<IntOrId>{i, i + 12});
     }
@@ -395,7 +395,7 @@ TEST(OptionalJoin, gallopingJoin) {
       bInput.emplace_back(std::vector<IntOrId>{i, i + 12});
     }
     auto numElementsInLarger = static_cast<int64_t>(
-        std::max(uint64_t{10000}, a.numRows() * GALLOP_THRESHOLD + 1));
+        std::max(size_t{10000}, a.numRows() * GALLOP_THRESHOLD + 1));
     for (int64_t i = 400; i < numElementsInLarger; ++i) {
       bInput.emplace_back(std::vector<IntOrId>{i, i + 12});
     }
